requirements.txt
streamlit==1.28.1
boto3==1.34.0
requests==2.31.0
Pillow==10.0.1
PyMuPDF==1.23.0